test: rename ArchUnit tests to IT so that it gets run by the CI
